+++ src/main/java/webserver/Controller.java
@@ -1,7 +1,5 @@
 package webserver;
 
-import webserver.httpResponse.HttpResponse;
-
 public interface Controller {
     String service(HttpRequest request);
 }
