+++ src/main/java/webserver/httpResponseProcessor/Http2xxResponseProcessor.java
@@ -1,8 +1,9 @@
-package webserver.httpResponse;
+package webserver.httpResponseProcessor;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
+import webserver.HttpResponse;
 import webserver.HttpResponseProcessor;
 
 import java.io.DataOutputStream;
