+++ src/main/java/webserver/viewProcessor/HtmlViewProcessor.java
@@ -1,6 +1,7 @@
 package webserver.viewProcessor;
 
-import webserver.httpResponse.HttpResponse;
+import webserver.HttpResponse;
+import webserver.ViewProcessor;
 
 public class HtmlViewProcessor implements ViewProcessor {
